+++ src/components/agent/controller.ts
@@ -21,16 +21,6 @@ export class AgentController {
     }
   }
 
-  // static async showPayments(_req: Req, res: Res, next: NextFn) {
-  //   try {
-  //     const payments = await AgentServices.showPayments();
-
-  //     res.status(OK).json(apiResponse(payments));
-  //   } catch (error) {
-  //     next(error);
-  //   }
-  // }
-
   static async releasePayment(req: Req, res: Res, next: NextFn) {
     try {
       const { id } = req.params;